notice controller api 전체 수정

+++ src/main/java/bucheon/leafy/application/service/NoticeService.java
@@ -1,16 +1,13 @@
 package bucheon.leafy.application.service;
 
-import bucheon.leafy.application.mapper.AlarmMapper;
 import bucheon.leafy.application.mapper.NoticeMapper;
-import bucheon.leafy.application.repository.UserRepository;
+import bucheon.leafy.config.AuthUser;
 import bucheon.leafy.domain.alarm.AlarmType;
 import bucheon.leafy.domain.notice.request.NoticeEditRequest;
 import bucheon.leafy.domain.notice.request.NoticeSaveRequest;
 import bucheon.leafy.domain.notice.response.NoticeEditResponse;
 import bucheon.leafy.domain.notice.response.NoticeResponse;
 import bucheon.leafy.domain.notice.response.NoticeSaveResponse;
-import bucheon.leafy.domain.user.User;
-import bucheon.leafy.domain.user.response.GetMeResponse;
 import bucheon.leafy.exception.*;
 import bucheon.leafy.util.request.PageRequest;
 import bucheon.leafy.util.response.PageResponse;
@@ -26,39 +23,42 @@ import java.util.List;
 @RequiredArgsConstructor
 public class NoticeService {
 
-    private final UserRepository userRepository;
     private final NoticeMapper noticeMapper;
     private final AlarmService alarmService;
-    private final AlarmMapper alarmMapper;
 
 
-    public int remove(Long noticeId) {
+    public void remove(Long noticeId, AuthUser user) {
+        NoticeResponse result = noticeMapper.findByIdAndIsDeleteFalse(noticeId);
+
+        if(result == null){ throw new NoticeNotFoundException(); }
+
+        if(result.getUserId() != user.getUserId()){ throw new NoticeNotUserIdException(); }
+
         if (noticeMapper.deleteById(noticeId) != 1) {
             throw new RemoveFailedException();
         }
-        return noticeMapper.deleteById(noticeId);
     }
 
 
-    public NoticeSaveResponse write(NoticeSaveRequest noticeSaveRequest , Long userId) {
+    public NoticeSaveResponse write(NoticeSaveRequest noticeSaveRequest, AuthUser user) {
+        Long userId = user.getUserId();
 
+        if (noticeMapper.save(userId, noticeSaveRequest) != 1) { throw new WriteFailedException(); }
 
-        if (noticeMapper.save(noticeSaveRequest) == null) {
-            throw new WriteFailedException();
-        }
-        NoticeSaveResponse noticeSaveResponse = noticeMapper.save(noticeSaveRequest);
+        NoticeResponse result = noticeMapper.findById(noticeSaveRequest.getNoticeId());
 
+        // 알림 발송
         List<Long> userIds = noticeMapper.findAllUserIds();
         for (Long id : userIds) {
             if (!id.equals(userId)) {
-                alarmService.createAlarm(id, AlarmType.NOTICE, noticeSaveResponse.getNoticeId());
+                alarmService.createAlarm(id, AlarmType.NOTICE, noticeSaveRequest.getNoticeId());
             }
         }
 
-        return noticeSaveResponse;
+        return NoticeSaveResponse.of(noticeSaveRequest.getNoticeId(), result.getCreatedAt());
     }
 
-
+    //TODO 어드민이랑 유저(비회원도)랑 리스트 다르게 뿌리기 (어드민은 isHide true 인것도 보여줘야함)
     public PageResponse getList(PageRequest pageRequest)  {
         List<NoticeResponse> list = noticeMapper.pageFindById(pageRequest);
         long total = noticeMapper.count();
@@ -67,37 +67,31 @@ public class NoticeService {
         return pageResponse;
     }
 
-    @Transactional
+    //TODO 어드민이랑 유저(비회원도)랑 리스트 다르게 뿌리기 (어드민은 isHide true 인것도 보여줘야함)
     public NoticeResponse getRead(Long noticeId) {
-
-        if (noticeMapper.findById(noticeId) == null) {
-            throw new ReadFailedException();
+        if (noticeMapper.findByIdAndIsDeleteFalseAndIsHideFalse(noticeId) == null) {
+            throw new NoticeNotFoundException();
         }
         noticeMapper.viewCnt(noticeId);
 
-        return noticeMapper.findById(noticeId);
+        return noticeMapper.findByIdAndIsDeleteFalseAndIsHideFalse(noticeId);
     }
-    public NoticeEditResponse modify(Long noticeId, NoticeEditRequest noticeEditRequest)  {
 
-        NoticeEditResponse noticeEditResponse = noticeMapper.editById(noticeId, noticeEditRequest);
+    public NoticeEditResponse modify(Long noticeId, NoticeEditRequest noticeEditRequest, AuthUser user)  {
 
-        if (noticeEditResponse == null) {
+        NoticeResponse result = noticeMapper.findByIdAndIsDeleteFalse(noticeId);
+
+        if(result == null){ throw new NoticeNotFoundException(); }
+
+        if(result.getUserId() != user.getUserId()){ throw new NoticeNotUserIdException(); }
+
+        if (noticeMapper.editById(noticeId, noticeEditRequest) != 1) {
             throw new ModifyFailedException();
         }
 
-        return noticeEditResponse;
+        NoticeResponse editResult = noticeMapper.findByIdAndIsDeleteFalse(noticeId);
+        return NoticeEditResponse.of(editResult.getModifiedAt());
     }
 
-    public void hideByNoticeId(Long noticeId){
-        noticeMapper.hideByNoticeId(noticeId);
-    };
-    public User getUserById(Long userId){
-        return userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
-    }
 
-    public GetMeResponse getMe(Long userId) {
-        User user = getUserById(userId);
-        int alarmCount = alarmMapper.countByUserId(userId);
-        return GetMeResponse.of(user, alarmCount);
-    }
 }
